+++ Tests/SwiftStdlibTests/ArrayExtensionsTests.swift
@@ -46,20 +46,6 @@ final class ArrayExtensionsTests: XCTestCase {
         XCTAssertEqual(swappedEmptyArray, emptyArray)
     }
 
-    func testDivided() {
-        let input = [0, 1, 2, 3, 4, 5]
-        let (even, odd) = input.divided { $0 % 2 == 0 }
-        XCTAssertEqual(even, [0, 2, 4])
-        XCTAssertEqual(odd, [1, 3, 5])
-
-        // Parameter names + indexes
-        let tuple = input.divided { $0 % 2 == 0 }
-        XCTAssertEqual(tuple.matching, [0, 2, 4])
-        XCTAssertEqual(tuple.0, [0, 2, 4])
-        XCTAssertEqual(tuple.nonMatching, [1, 3, 5])
-        XCTAssertEqual(tuple.1, [1, 3, 5])
-    }
-
     func testKeyPathSorted() {
         let array = [Person(name: "James", age: 32), Person(name: "Wade", age: 36), Person(name: "Rose", age: 29)]